// frontend/src/api/generated/models/ImpactDraftDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ImpactDraftDTO = {
    event_id: number;
    title: string;
    t0?: (string | null);
    computed_at?: (string | null);
    assets?: Record<string, Record<string, any>>;
};